chore: Added type defintions for the DiscoverPage template
